* src/itdb_itunesdb.c,itdb_private.h: fixes for 64 bit systems 	* src/ipod-device.c: fixes for 64 bit systems (patch provided by 	  Bodo Bauer) 	* TROUBLESHOOTING: hint for 64 bit compilation
git-svn-id: https://gtkpod.svn.sourceforge.net/svnroot/gtkpod/libgpod/trunk@1121 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/ipod-device.c
@@ -1733,9 +1733,9 @@ ipod_device_debug(IpodDevice *device)
 	g_printf("   Serial Number:    %s\n", serial_number);
 	g_printf("   Firmware Version: %s\n", firmware_version);
 	g_printf("Volume Info\n");
-	g_printf("   Volume Size:      %lld\n", volume_size);
-	g_printf("   Volume Used:      %lld\n", volume_used);
-	g_printf("   Available         %lld\n", volume_available);
+	g_printf("   Volume Size:      %lld\n", (long long int)volume_size);
+	g_printf("   Volume Used:      %lld\n", (long long int)volume_used);
+	g_printf("   Available         %lld\n", (long long int)volume_available);
 	g_printf("   UUID:             %s\n", volume_uuid);
 	g_printf("   Label             %s\n", volume_label);
 	g_printf("User-Provided Info\n");
